--- backend/app/Http/Controllers/AuthController.php
@@ -1,93 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-use App\Models\User;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Cookie;
-use Illuminate\Support\Facades\Hash;
-use Symfony\Component\HttpFoundation\Response;
-
-class AuthController extends Controller
-
-{
-   
-
-        // public function register(Request $request)
-        // {
-        //     try{
-        //     return User::create([
-        //         'fullname' => $request->input('fullname'),
-        //         'email' => $request->input('email'),
-        //         'password' => Hash::make($request->input('password'))
-        //     ]);
-        // }catch(\Exception $err){
-        //     return response()->json(['error' => $err,] , 404); 
-
-        // }
-        // }
-
-
-        public function register(Request $request ){
-            try {
-    
-            $user = new User;
-            
-        
-            $fullname = $request->input('fullname');
-            $email = $request->input('email');
-            $password = Hash::make($request->input('password'));
-    
-            $user->fullname = $fullname;
-            $user->email = $email;
-            $user->password = $password;
-    
-            $user->save();
-    
-            return response()->json([
-                'message' => 'user created successfully!',
-            ]);
-    
-        }catch (\Exception $err) {
-            return response()->json(['message' => 'Error adding user: '. $err->getMessage(),], 500); 
-        }
-    }
-
-
-    
-
-        public function login(Request $request)
-        {
-            if (!Auth::attempt($request->only('email', 'password'))) {
-                return response([
-                    'message' => 'Invalid credentials!'
-                ], Response::HTTP_UNAUTHORIZED);
-            }
-    
-            $user = Auth::user();
-    
-            $token = $user->createToken('token')->plainTextToken;
-    
-            $cookie = cookie('jwt', $token, 60 * 24); // 1 day
-    
-            return response([
-                'message' => $token
-            ])->withCookie($cookie);
-        }
-    
-        public function user()
-        {
-            return Auth::user();
-        }
-    
-        public function logout()
-        {
-            $cookie = Cookie::forget('jwt');
-    
-            return response([
-                'message' => 'Success'
-            ])->withCookie($cookie);
-        }
-
-}
