+++ backend/app/Http/Middleware/Authenticate.php
@@ -12,7 +12,7 @@ class Authenticate extends Middleware
      *
      * @param \Illuminate\Http\Request $request
      * @return string|null
-     */
+     */   
     protected function redirectTo($request)
     {
         if (!$request->expectsJson()) {
